+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Change Log
 ==========
 
+## Versopm 2.0.1
+### New
+- Added Redact annotation for value parameters which should not be passed to krang
+- Added Gradle extension and CLI parameter to disable krang during compilation
+- Added runtime flag to disable krang during runtime
 ## Version 2.0.0
 ### New
 - Made the runtime library multiplatform